feat:fetching videos from appwrite &  custom video compoents
