+++ app/common/post-types.ts
@@ -1,0 +1,5 @@
+export interface PostItem {
+  $id: string;
+  video: string;
+  thumbnail: string;
+}
