+++ components/TrendingItem.tsx
@@ -1,0 +1,61 @@
+import React, { useState } from "react";
+import { ResizeMode, Video, AVPlaybackStatus } from "expo-av";
+import * as Animatable from "react-native-animatable";
+import { Image, ImageBackground, TouchableOpacity } from "react-native";
+import { icons } from "../constants";
+import { PostItem } from "@/app/common/post-types";
+import { zoomIn, zoomOut } from "@/utils/animation";
+
+interface TrendingItemProps {
+  activeItem: string;
+  item: PostItem;
+}
+
+const TrendingItem: React.FC<TrendingItemProps> = ({ activeItem, item }) => {
+  const [play, setPlay] = useState<boolean>(false);
+
+  return (
+    <Animatable.View
+      className="mr-5"
+      animation={activeItem === item.$id ? zoomIn : zoomOut}
+      duration={500}
+    >
+      {play ? (
+        <Video
+          source={{ uri: item.video }}
+          className="w-52 h-72 rounded-[33px] mt-3 bg-white/10"
+          resizeMode={ResizeMode.CONTAIN}
+          useNativeControls
+          shouldPlay
+          onPlaybackStatusUpdate={(status: AVPlaybackStatus) => {
+            if (status.isLoaded && status.didJustFinish) {
+              setPlay(false);
+            }
+          }}
+        />
+      ) : (
+        <TouchableOpacity
+          className="relative flex justify-center items-center"
+          activeOpacity={0.7}
+          onPress={() => setPlay(true)}
+        >
+          <ImageBackground
+            source={{
+              uri: item.thumbnail,
+            }}
+            className="w-52 h-72 rounded-[33px] my-5 overflow-hidden shadow-lg shadow-black/40"
+            resizeMode="cover"
+          />
+
+          <Image
+            source={icons.play}
+            className="w-12 h-12 absolute"
+            resizeMode="contain"
+          />
+        </TouchableOpacity>
+      )}
+    </Animatable.View>
+  );
+};
+
+export default TrendingItem;
