+++ utils/animation.ts
@@ -1,0 +1,25 @@
+export const zoomIn = {
+  0: {
+    scale: 0.9,
+  },
+  0.5: {
+    scale: 0.95,
+  },
+  1: {
+    scale: 1,
+  },
+  easing: 'ease-out',
+};
+
+ export const zoomOut = {
+  0: {
+    scale: 1,
+  },
+  0.5: {
+    scale: 0.95,
+  },
+  1: {
+    scale: 0.9,
+  },
+  easing: 'ease-in',
+};
